check status of megacli volumes, patch from Steffen Zieger <me#saz.sh>:
Hello glen,

check_raid is not checking, if a volume is degraded on LSI SAS
Controllers.
It only checks through MegaCLI, if there is a failed drive,
but as soon as this drive is removed, everythings back to normal.

This patch will fix it (at least for me).

Changed files:
    check_raid -> 1.99