3ware: VERIFYING is a normal operation
The code of check_raid contains an obvious bug in trying
to do a string-compare on a regexp. In

        /usr/lib/nagios/plugins/check_raid
        check_tw_cli(), line 738:

        } elsif ($s eq 'INITIALIZING|VERIFYING') {

The effect is that both status (INITIALIZING and VERIFYING)
will not be recognized and be reported as UNKNOWN. Also
the VERIFYING state will happen as part of a normal operation.
That is especially true if the auto-verification feature
is enabled. To avoid noise this state should be classified
as OK.

Patch was submitted by Harald Wilhelmi <harald.wilhelmi@tngtech.com>
in Debian Bug #689297 (http://bugs.debian.org/689297)