Fix false positive state on HP controllers
On HP Raid-Controllers were just one disk has failed (and therefore the raid itself is working and ok) the output of hpacucli is NOT this:

array A (Failed)

but just:

array A

Therefore the check assumes everything is ok and doesn't do the extra details check. Since we can not really nicely fix this: $array{$array}[0] = $s || 'OK';
I just removed the latter check for the 'OK' where it only does the extra details check if it is not 'OK'.
Only disadvantage is that everytime the extra details checks are done and displayed but that does not hurt as much as an unrecognized failed raid.