do not report critical on hot-spare.
check_raid pushes a critical state for any physical disk whose status is
not reported as "Online". However, Hot-Spares are reported as such in
the status field of arcconf's getconfig output, even though everything
is fine and the spares are not in use. This commit allows for this.

Core functionality should not be affected, as any degradation (and hence
activation of a hot spare) will be reported in other areas of arccconf's
output, and cause check_raid to report an appropiate check state.