cciss: allow cciss_vol_status version with letter
The latest version of cciss_vol_status is 1.12a. For the first time, an
extra letter was appended.

The version pattern would not match any more. The function will return
0, and later on, the code to detect if the version is >= 1.10 (for new
added functionality) would wrongly return false.

So, change it to also accept an optional letter at the end, but ignore
it. Adding it to the detected version string will result in errors like
"Argument "1.12a" isn't numeric" later on in the code, and because right
now, there is no need to be able to distinguish between 1.12 and 1.12a
in the plugin.