cciss: adapt for changed physical drive output
In version 1.12a, cciss_vol_status changes the output of physical location of drives:

    -       sprintf(tail, " connector %c%c box %d bay %d %40s %40s %8s",
    +       sprintf(tail, " connector %c%c box %d bay %-2d %40s %40s %8s",

This means that the bay number is now 2 posititions, left aligned.
So, e.g. '1 ', '6 ', or '12'.

Allow for an extra space to be present when the bay number is less than
10.

Without this change, the pattern will not match any more, and the
program complains about Unparsed lines.