Test data for lsraid
The lsraid plugin is marked broken because of missing test data. This commit adds lsraid output for centos2 and centos3 hosts.
The plugin is seemingly broken though because it expects a single space instead of multiple spaces inside the regex. It also will ignore missing disks as they do not match the regex.